+++ apps/client/src/app/layout.tsx
@@ -27,7 +27,7 @@ export default function RootLayout({ children }: { children: ReactNode }) {
       <body className={quickSand.className}>
         <Providers>
           <Header />
-          <main className="min-h-screen w-full bg-stone-50">
+          <main className="min-h-screen w-full bg-slate-50">
             {children}
           </main>
         </Providers>